requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
reportlab==4.0.7
schedule==1.2.0
pyserial==3.5
